refactor(tarko-ui): reorganize components with PascalCase directory naming
