+++ multimodal/tarko/ui/src/components/Tooltip/index.ts
@@ -1,0 +1,1 @@
+export { Tooltip } from './Tooltip';
