+++ multimodal/tarko/ui/src/components/dialog/index.ts
@@ -1,1 +1,2 @@
+export { Dialog, DialogPanel, DialogTitle } from './Dialog';
 export * from './ConfirmDialog';
