--- multimodal/tarko/ui/src/components/mui/index.ts
@@ -1,17 +1,0 @@
-export { Dialog, DialogPanel, DialogTitle } from './MuiDialog';
-export * from './Tooltip';
-
-// Direct re-exports from @mui/material
-export {
-  Box,
-  Typography,
-  CircularProgress,
-  Select,
-  MenuItem,
-  FormControl,
-  Menu,
-  IconButton,
-  Divider
-} from '@mui/material';
-
-export { ThemeProvider } from '@mui/material/styles';
